+++ odoo/local-src/specific_sale/models/sale.py
@@ -571,3 +571,16 @@ class SaleOrderLine(models.Model):
             if self.order_id.partner_id.alcyon_category_id == group:
                 return True
         return False
+
+    def validate_no_psychotropic_ordered_by_phone(self):
+        """No psychotropic ordered on the phone."""
+        psycho_cat = self.env.ref('specific_data.product_categ_stupefiant')
+        if not self.product_id.categ_id.has_for_parent(psycho_cat.id):
+            return False
+        return self.order_id.sale_channel == 'phone'
+
+    # Warnings
+    def warning_psychotropic(self):
+        """Add warning for psychotropic product on sale order line."""
+        psycho_cat = self.env.ref('specific_data.product_categ_stupefiant')
+        return self.product_id.categ_id.has_for_parent(psycho_cat.id)